+++ features/blog/react-feature-blog/src/feature.tsx
@@ -1,5 +1,7 @@
+import { BlogQueryConfig } from '@/content/blog-post/blog-query-config';
 import { FeatureDescriptor } from '@vyuh/react-core';
 import { ContentExtensionDescriptor } from '@vyuh/react-extension-content';
+import { DocumentViewDescriptor } from '@vyuh/react-feature-system';
 import { FaPagelines as Icon } from 'react-icons/fa6';
 import { BlogAuthorContentBuilder } from './content/blog-author';
 import { BlogCategoryContentBuilder } from './content/blog-category';
@@ -14,6 +16,11 @@ export const blog = new FeatureDescriptor({
   icon: <Icon />,
   extensions: [
     new ContentExtensionDescriptor({
+      contents: [
+        new DocumentViewDescriptor({
+          queries: [BlogQueryConfig.typeDescriptor],
+        }),
+      ],
       contentBuilders: [
         new BlogGroupContentBuilder(),
         new BlogPostSummaryContentBuilder(),