+++ features/auth/react-feature-auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vyuh/react-feature-auth
 
+## 0.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @vyuh/react-feature-system@0.6.0
+  - @vyuh/react-core@0.6.0
+  - @vyuh/react-extension-content@0.6.0
+
 ## 0.5.9
 
 ### Patch Changes